fix: set base fee in chainspec

+++ crates/node/src/node.rs
@@ -362,7 +362,7 @@ where
 
     async fn build_pool(self, ctx: &BuilderContext<Node>) -> eyre::Result<Self::Pool> {
         let mut pool_config = ctx.pool_config().clone();
-        pool_config.minimal_protocol_basefee = TEMPO_BASE_FEE as u64;
+        pool_config.minimal_protocol_basefee = TEMPO_BASE_FEE;
 
         let blob_cache_size = if let Some(blob_cache_size) = pool_config.blob_cache_size {
             Some(blob_cache_size)
