+++ crates/node/tests/it/eth_call.rs
@@ -72,7 +72,7 @@ async fn test_eth_call() -> eyre::Result<()> {
     let mint_amount = U256::random();
     token
         .mint(caller, mint_amount)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -144,7 +144,7 @@ async fn test_eth_trace_call() -> eyre::Result<()> {
     let mint_amount = U256::random();
     token
         .mint(caller, mint_amount)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -258,7 +258,7 @@ async fn test_eth_get_logs() -> eyre::Result<()> {
     let mint_amount = U256::random();
     let mint_receipt = token
         .mint(caller, mint_amount)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -268,7 +268,7 @@ async fn test_eth_get_logs() -> eyre::Result<()> {
     let recipient = Address::random();
     token
         .transfer(recipient, mint_amount)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
