+++ crates/node/tests/it/tip20.rs
@@ -84,7 +84,7 @@ async fn test_tip20_transfer() -> eyre::Result<()> {
         pending_txs.push(
             token
                 .mint(*account, *balance)
-                .gas_price(TEMPO_BASE_FEE)
+                .gas_price(TEMPO_BASE_FEE as u128)
                 .gas(30000)
                 .send()
                 .await?,
@@ -118,7 +118,7 @@ async fn test_tip20_transfer() -> eyre::Result<()> {
         assert!(success);
         let pending_tx = token
             .transfer(recipient, sender_balance)
-            .gas_price(TEMPO_BASE_FEE)
+            .gas_price(TEMPO_BASE_FEE as u128)
             .gas(30000)
             .send()
             .await?;
@@ -212,7 +212,7 @@ async fn test_tip20_mint() -> eyre::Result<()> {
         pending_txs.push(
             token
                 .mint(*account, *balance)
-                .gas_price(TEMPO_BASE_FEE)
+                .gas_price(TEMPO_BASE_FEE as u128)
                 .gas(30000)
                 .send()
                 .await?,
@@ -309,7 +309,7 @@ async fn test_tip20_transfer_from() -> eyre::Result<()> {
     let total_balance: U256 = account_data.iter().map(|(_, balance)| *balance).sum();
     token
         .mint(caller, total_balance)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -324,7 +324,7 @@ async fn test_tip20_transfer_from() -> eyre::Result<()> {
         pending_txs.push(
             token
                 .approve(signer.address(), *balance)
-                .gas_price(TEMPO_BASE_FEE)
+                .gas_price(TEMPO_BASE_FEE as u128)
                 .gas(30000)
                 .send()
                 .await?,
@@ -364,7 +364,7 @@ async fn test_tip20_transfer_from() -> eyre::Result<()> {
 
         let pending_tx = spender_token
             .transferFrom(caller, recipient, *allowance)
-            .gas_price(TEMPO_BASE_FEE)
+            .gas_price(TEMPO_BASE_FEE as u128)
             .gas(30000)
             .send()
             .await?;
@@ -432,7 +432,7 @@ async fn test_tip20_transfer_with_memo() -> eyre::Result<()> {
     let recipient = Address::random();
     token
         .mint(caller, transfer_amount)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -443,7 +443,7 @@ async fn test_tip20_transfer_with_memo() -> eyre::Result<()> {
     let memo = FixedBytes::<32>::random();
     let receipt = token
         .transferWithMemo(recipient, transfer_amount, memo)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -515,7 +515,7 @@ async fn test_tip20_blacklist() -> eyre::Result<()> {
     // Create a blacklist policy
     let policy_receipt = registry
         .createPolicy(admin, ITIP403Registry::PolicyType::BLACKLIST)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -533,7 +533,7 @@ async fn test_tip20_blacklist() -> eyre::Result<()> {
     // Update the token policy to the blacklist
     token
         .changeTransferPolicyId(policy_id)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -557,7 +557,7 @@ async fn test_tip20_blacklist() -> eyre::Result<()> {
     try_join_all(blacklisted_accounts.iter().map(|account| async {
         registry
             .modifyPolicyBlacklist(policy_id, account.address(), true)
-            .gas_price(TEMPO_BASE_FEE)
+            .gas_price(TEMPO_BASE_FEE as u128)
             .gas(30000)
             .send()
             .await
@@ -571,7 +571,7 @@ async fn test_tip20_blacklist() -> eyre::Result<()> {
     try_join_all(accounts.iter().map(|account| async {
         token
             .mint(account.address(), U256::from(1000))
-            .gas_price(TEMPO_BASE_FEE)
+            .gas_price(TEMPO_BASE_FEE as u128)
             .gas(30000)
             .send()
             .await
@@ -611,7 +611,7 @@ async fn test_tip20_blacklist() -> eyre::Result<()> {
 
             token
                 .transfer(Address::random(), U256::ONE)
-                .gas_price(TEMPO_BASE_FEE)
+                .gas_price(TEMPO_BASE_FEE as u128)
                 .gas(30000)
                 .send()
                 .await
@@ -670,7 +670,7 @@ async fn test_tip20_whitelist() -> eyre::Result<()> {
     // Create a whitelist policy
     let policy_receipt = registry
         .createPolicy(admin, ITIP403Registry::PolicyType::WHITELIST)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -688,7 +688,7 @@ async fn test_tip20_whitelist() -> eyre::Result<()> {
     // Update the token policy to the whitelist
     token
         .changeTransferPolicyId(policy_id)
-        .gas_price(TEMPO_BASE_FEE)
+        .gas_price(TEMPO_BASE_FEE as u128)
         .gas(30000)
         .send()
         .await?
@@ -721,7 +721,7 @@ async fn test_tip20_whitelist() -> eyre::Result<()> {
     try_join_all(whitelisted_accounts.iter().map(|account| async {
         registry
             .modifyPolicyWhitelist(policy_id, *account, true)
-            .gas_price(TEMPO_BASE_FEE)
+            .gas_price(TEMPO_BASE_FEE as u128)
             .gas(30000)
             .send()
             .await
@@ -735,7 +735,7 @@ async fn test_tip20_whitelist() -> eyre::Result<()> {
     try_join_all(accounts.iter().map(|account| async {
         token
             .mint(account.address(), U256::from(1000))
-            .gas_price(TEMPO_BASE_FEE)
+            .gas_price(TEMPO_BASE_FEE as u128)
             .gas(30000)
             .send()
             .await
@@ -782,7 +782,7 @@ async fn test_tip20_whitelist() -> eyre::Result<()> {
             .map(|(token, recipient)| async {
                 token
                     .transfer(*recipient, U256::ONE)
-                    .gas_price(TEMPO_BASE_FEE)
+                    .gas_price(TEMPO_BASE_FEE as u128)
                     .gas(30000)
                     .send()
                     .await
