+++ crates/node/tests/it/tip20_gas_fees.rs
@@ -66,7 +66,7 @@ async fn test_fee_in_stable() -> eyre::Result<()> {
     let tx = TransactionRequest::default()
         .from(caller)
         .to(caller)
-        .gas_price(TEMPO_BASE_FEE);
+        .gas_price(TEMPO_BASE_FEE as u128);
 
     let pending_tx = provider.send_transaction(tx).await?;
     let receipt = pending_tx.get_receipt().await?;
